Added project stage 4. Added GUI so far.
Decided to use Netbeans because it is much simpler with the ant build. I added the classes that we had for stage 3 and the new frame class that creates the JFrame. I just uploaded all of these files through the website; I didn't have to clone into netbeans. In netbeans I created a new ant build project and just copied/pasted the files from github project over to that project.